`defmt-parser`: Move `DisplayHint` to separate file
